Renamed Log to ActivityLog, fixed some import errors
