+++ src/model/__init__.py
@@ -1,7 +1,7 @@
 from Course import *
 from Group import *
 from Instructor import *
-from Log import *
+from ActivityLog import *
 from LogEntry import *
 from Response import *
 from Round import *
